feat(console): making render px by px

+++ src/sys/debug/debug.h
@@ -6,11 +6,11 @@
 
 extern void sys_debug_info_at(const u8* str, u8 x, u8 y);
 extern void sys_debug_char_at(const u16 ascii, u8 x, u8 y);
-extern void sys_debug_number_at(u8 num, u8 x, u8 y);
+extern void sys_debug_number_at(u16 num, u8 x, u8 y);
 
 extern void sys_debug_info(const u8* str);
 extern void sys_debug_char(const u16 ascii);
-extern void sys_debug_number(u8 num);
+extern void sys_debug_number(u16 num);
 
 extern void sys_debug_waitKey(void);
